==> 4_analysis/2a_dig_into_fe_fishing_pacific.Rmd <==
---
title: "Examine area of high FE non-cc impacts in Pacific"
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    code_folding: hide
    toc: true
    toc_depth: 3
    toc_float: yes
    number_sections: true
    theme: cerulean
    highlight: haddock
    includes: 
      in_header: '~/github/src/templates/ohara_hdr.html'
  pdf_document:
    toc: true
---

``` {r setup, echo = TRUE, message = FALSE, warning = FALSE}

knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE, fig.height = 4, fig.width = 7)

library(terra)
library(oharac)
library(data.table)
library(tidyverse)
library(here)
source(here('common_fxns.R'))
```

# Summary

There is a region between the coast of North America and Hawaii where FE non-cc impacts are noticeably higher than spp impacts.  Find this region and dig into why!

# Methods

- Identify these cells
- Generate a species list for these cells (and hab list?)
- calculate FV across all spp in the cell
- candidate species to focus on might be:
    * local high FV: species is the lone rep of an FE locally (but in an FE with multiple spp)
        * species FE list contains info on FE overall membership; compare to local species list in candidate area
    * regional moderate/low FV: species has a large range, such that it overlaps with other members of its FE in other regions
        * info on spp ranges is available
        * for candidate FE, generate a map of FE richness (or FV)
    * highly vulnerable to stressors that are patchy, creating local high impacts in some areas but low in others nearby

## Identify the cells

```{r}
fe_rasts <- list.files(here('_output/cumulative_impact_maps'), 
                       pattern = 'funct_entity_group', full.names = TRUE)
spp_rasts <- list.files(here('_output/cumulative_impact_maps'), 
                       pattern = 'species_group', full.names = TRUE)

fishing_rast <- rast(fe_rasts[str_detect(fe_rasts, 'fishing.tif')])
fe_noncc  <- rast(fe_rasts[!str_detect(fe_rasts, 'climate.tif')]) %>%
  sum(na.rm = TRUE)
spp_noncc <- rast(spp_rasts[!str_detect(spp_rasts, 'climate.tif')]) %>%
  sum(na.rm = TRUE)

cell_id <- fe_noncc %>%
  setValues(1:ncell(.))

diff_r <- fe_noncc - spp_noncc

diff_r[diff_r < 0.05 | diff_r > 0.11] <- NA

bounds <- ext(-1.5e7, -1e7, 2e6, 4e6)

cell_id_hi <- cell_id %>%
  mask(diff_r) %>%
  crop(bounds)
plot(cell_id_hi)

cell_df <- as.data.frame(cell_id_hi, xy = TRUE) %>%
  setNames(c('x', 'y', 'cell_id'))
```


### Create dataframe of species inhabiting these cells

Borrow code from the FV calculations, that clip species ranges to just those found in a given set of cells.

```{r define truncated rangemap functions}
read_cellvec_rangemap <- function(f, cell_vec) {
  # f <- spp_map_fs$map_f[1]
  if(file.exists(f)) {
    df <- data.table::fread(f) %>%
      filter(cell_id %in% cell_vec) %>%
      mutate(map_f = f)
    ### filter for presence and prob as needed
    if(str_detect(basename(f), '^am')) {
      df <- df %>%
        filter(prob >= .5) %>%
        select(-prob)
    } else {
      df <- df %>%
        filter(presence != 5) %>%
        select(-presence)
    }
  } else {
    df <- data.frame()
  }  
  if(nrow(df) == 0) {
    return(NULL) 
  } else {
    return(df %>% distinct())
  }
}

bind_maps_list <- function(maps_list, spp_for_calc) {
  ### NOTE: for some reason, the bind_rows() in here sometimes causes
  ### unrecoverable errors when knitting, but seems OK when running chunks
  ### individually... try subbing with data.table::rbindlist 
  if(check_tryerror(maps_list)) {
    stop('Try-errors detected in bind_maps_list()!')
  }
  maps_bound <- maps_list %>%
    ### drop NULL instances (no spp cells - helps keep things from crashing)
    purrr::compact() %>% 
    data.table::rbindlist()
  
  if(nrow(maps_bound) > 0) {
    ### if no spp-cell data for this chunk, skip bind and return 0-length df
    dt1 <- data.table::data.table(maps_bound, key = 'map_f')
    dt2 <- data.table::data.table(spp_for_calc, key = 'map_f')
    maps_bound <- merge(dt1, dt2, all.x = TRUE, allow.cartesian = TRUE) %>%
      as.data.frame() %>%
      select(-map_f) %>%
      distinct()
  }
  return(maps_bound)
}
```

Identify the species inhabiting this subset of marine ecoregion provinces.

```{r}
spp_info_df <- assemble_spp_info_df() %>%
  rename(vulnerability = stressor) %>%
  select('species', 'taxon', 'vulnerability', 'v_score',
         'fe_id', 'mob', 'trp', 'len', 'wcol',
         'src', 'id', 'map_f')

spp_map_fs <- spp_info_df %>%
  select(species, map_f) %>%
  group_by(map_f) %>%
  filter(species == first(species)) %>%
  distinct()
spp_fe <- spp_info_df %>%
  select(species, fe_id, mob:wcol, taxon) %>%
  distinct() %>%
  group_by(fe_id) %>%
  mutate(nspp_fe = n_distinct(species)) %>%
  ungroup()

cell_vec <- cell_df$cell_id[1:100] %>% unique() %>% sort()

patch_spp_list <- parallel::mclapply(spp_map_fs$map_f, cell_vec = cell_vec,
                                     mc.cores = 40,
                                     FUN = read_cellvec_rangemap)

patch_spp_df <- bind_maps_list(patch_spp_list, spp_map_fs)

spp_vec <- unique(patch_spp_df$species)
```

### Gather biomass removal stressor maps

```{r collect catch maps function}
collect_catch_maps <- function(catch_mapfile_df, catch_cells) {

  ### read in all biomass removal stressor maps for select species
  catch_maps_list <- parallel::mclapply(
          catch_mapfile_df$catch_map_f, ### f <- catch_mapfile_df$catch_map_f[27]
          mc.cores = 40, cells = catch_cells,
          FUN = function(f, cells) {
            df <- data.table::fread(f) 
            if(nrow(df) > 0) {
              df <- df %>% 
                mutate(rescaled_catch = as.numeric(rescaled_catch)) %>%
                filter(cell_id %in% cells)
            }
            return(df)
          })

  if(check_tryerror(catch_maps_list)) {
    message('Something went wrong with loading maps!')
  }
  message('Binding biomass removal stress maps...')
  catch_maps <- catch_maps_list %>%
    setNames(catch_mapfile_df$species) %>%
    data.table::rbindlist(idcol = 'species')
  
  return(catch_maps)
}
```

``` {r}
catch_map_stem <- here_anx('stressors/fishing/4_rescaled_catch_by_spp_cell', 
                           '%s_spp_rescaled_catch_%s.csv')

catch_mapfile_df <- spp_info_df %>%
  mutate(catch_map_f = sprintf(catch_map_stem, src, id)) %>%
  select(species, src, catch_map_f) %>%
  filter(file.exists(catch_map_f)) %>%
  distinct() %>%
  filter(species %in% spp_vec)

spp_catch_vuln_df <- spp_info_df %>%
  filter(vulnerability == 'biomass_removal')
spp_catch_maps <- collect_catch_maps(catch_mapfile_df, catch_cells = cell_vec) %>%
  left_join(spp_catch_vuln_df, by = 'species') %>%
  select(species, cell_id, vulnerability, v_score, intensity = rescaled_catch) %>%
  mutate(impact = ifelse(!is.na(intensity), v_score * intensity, 0))
  
```

## Examine regions

```{r examine fishing impacts}

patch_cells <- patch_spp_df %>%
  group_by_at(vars(-species)) %>%
  summarize(n_spp = n_distinct(species), .groups = 'drop')

spp_df <- patch_spp_df %>%
  left_join(spp_fe, by = 'species') %>%
  group_by(cell_id) %>%
  mutate(n_spp = n_distinct(species)) %>%
  group_by(cell_id, fe_id) %>%
  mutate(nspp_fe_present = n_distinct(species)) %>%
  ungroup() %>%
  left_join(spp_catch_maps)

check_spp_df <- spp_df %>%
  filter(nspp_fe_present < 4 & impact > .4)


```

Looks like lots of cells of skipjack, _Katsuwonus pelamis_, FV = 1, with high impacts

### Map skipjack impacts

```{r}
skipjack_richness_check <- spp_df %>%
  group_by(cell_id) %>%
  filter('katsuwonus pelamis' %in% species) %>%
  summarize(n_spp = n(),
            n_fe = n_distinct(fe_id))
  
skipjack_catch <- catch_mapfile_df %>%
  filter(species == 'katsuwonus pelamis') %>%
  pull(catch_map_f) %>%
  read_csv() %>%
  left_join(cell_df, by = 'cell_id') %>%
  filter(x > bounds[1] & x < bounds[2] & y > bounds[3] & y < bounds[4])

skipjack_impact <- skipjack_catch %>%
  mutate(species = 'katsuwonus pelamis') %>%
  left_join(spp_catch_vuln_df, by = 'species') %>%
  select(x, y, vulnerability, v_score, intensity = rescaled_catch) %>%
  mutate(impact = ifelse(!is.na(intensity), v_score * intensity, 0))

ggplot(skipjack_impact, aes(x, y, fill = impact)) +
  geom_raster() +
  scale_fill_viridis_c() +
  coord_fixed()
```

That looks pretty much like the exact pattern of skipjack fishing.
