==> 4_analysis/3_calc_taxa_representation.Rmd <==
---
title: "Calculate taxonomic representation"
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    code_folding: hide
    toc: true
    toc_depth: 3
    toc_float: yes
    number_sections: true
    theme: cerulean
    highlight: haddock
    includes: 
      in_header: '~/github/src/templates/ohara_hdr.html'
  pdf_document:
    toc: true
---

``` {r setup, echo = TRUE, message = FALSE, warning = FALSE}

knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE, fig.height = 4, fig.width = 7)

library(terra)
library(oharac)
library(data.table)
library(tidyverse)
library(here)
source(here('common_fxns.R'))
```

# Summary

What proportion of each taxon (based on totals in WoRMS) is represented in this analysis?

# Methods

```{r}
w <- assemble_worms()
x <- assemble_spp_info_df() %>%
  select(species, taxon) %>%
  distinct()

df <- w %>%
  mutate(repr = species %in% x$species)

phylum_df <- df %>%
  group_by(phylum, class) %>%
  summarize(n_tot = n_distinct(species),
            n_repr = sum(repr),
            p_repr = round(n_repr / n_tot, 3)) %>%
  arrange(desc(p_repr))

chord_df <- df %>%
  filter(phylum == 'chordata') %>%
  group_by(class, order) %>%
  summarize(n_tot = n_distinct(species),
            n_repr = sum(repr),
            p_repr = round(n_repr / n_tot, 3)) %>%
  arrange(desc(p_repr))

cnid_df <- df %>%
  filter(phylum == 'cnidaria') %>%
  group_by(class, order) %>%
  summarize(n_tot = n_distinct(species),
            n_repr = sum(repr),
            p_repr = round(n_repr / n_tot, 3)) %>%
  arrange(desc(p_repr))

DT::datatable(phylum_df)

DT::datatable(class_df)

mamm_df <- df %>%
  filter(class == 'mammalia') %>%
  group_by(order, family) %>%
  summarize(n_tot = n_distinct(species),
            n_repr = sum(repr),
            p_repr = round(n_repr / n_tot, 3)) %>%
  arrange(desc(p_repr))
  
carn_df <- df %>%
  filter(order == 'carnivora') %>%
  group_by(family) %>%
  summarize(n_tot = n_distinct(species),
            n_repr = sum(repr),
            p_repr = round(n_repr / n_tot, 3)) %>%
  arrange(desc(p_repr))
  
```
