==> 4_analysis/5b_spp_vs_fv_sens_analysis_impact_summary.Rmd <==
---
title: 'Sensitivity analysis: species-weighting vs FV weighting'
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    code_folding: hide
    toc: true
    toc_depth: 3
    toc_float: yes
    number_sections: true
    theme: cerulean
    highlight: haddock
    includes: 
      in_header: '~/github/src/templates/ohara_hdr.html'
  pdf_document:
    toc: true
---

``` {r setup, echo = TRUE, message = FALSE, warning = FALSE}

knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE,
                      fig.height = 4, fig.width = 7)

library(terra)
library(oharac)
library(data.table)
library(tidyverse)
library(here)
source(here('common_fxns.R'))
```

# Summary

Sensitivity analysis to determine how drastically scores change when remixing species and functional entities.  

In script 5a, we calculated sensitivity around species communities and per-stressor impacts.  In this script, we aggregate results across all stressors for each randomized iteration, to determine the distribution of mean values for cumulative impacts.

For each cell, across all iterations, we calculate a difference between the FE approach and spp approach, and then identify the 97.5% and 2.5% values across all iterations to calculate a confidence interval.  Note that for the difference calculation, a positive value indicates that the FE approach estimated a HIGHER impact than the SPP approach.

# Methods

For each randomized species set, collect iterations across all stressors and calculate cumulative impact for each iteration.  Because each iteration used the same seed for the pseudorandom number generator, the randomized species set should be identical across all stressors.  This allows us to simply add all stressors cell-by-cell for a given iteration to determine the cumulative impact.

Unfortunately, the tmp impact files for uniform-exposure stressors look like they're actually just vulnerability scores (impacts were calculated after).  So, for those stressors, bind with stressor intensity and convert to impacts... :/

``` {r select sample of cells}
spp_names_f <- here_anx('sens_analysis/spp_name_to_id.csv')
spp_cells_f <- here_anx('sens_analysis/spp_cells_to_id.csv')

if(any(!file.exists(spp_names_f, spp_cells_f))) {
  stop('Species/cell sample file(s) do not exist... Please run script 4a first to generate these files!')
}

spp_cells_sample_df <- read_csv(spp_names_f, show_col_types = FALSE) %>%
  left_join(read_csv(spp_cells_f, show_col_types = FALSE), by = 'spp_id') %>%
  select(-spp_id)
cell_id_vec <- spp_cells_sample_df$cell_id %>% unique() %>% sort()
n_cells <- length(cell_id_vec)
```

```{r load stressor intensities}
cell_id_r <- rast(here('_output/cumulative_impact_maps/chi_funct_entity.tif')) %>%
  setValues(1:ncell(.)) %>% setNames('cell_id')

stressor_intensity_f <- here_anx('sens_analysis/sample_str_intensity.csv')
if(!file.exists(stressor_intensity_f)) {
  
  str_vuln_lookup <- read_csv(here('_raw/stressor_vulnerability_lookup.csv'), 
                              show_col_types = FALSE)
  strs <- str_vuln_lookup$stressor
  strs_unif <- strs[!str_detect(strs, 'sst_rise|biomass_removal|bycatch')]
  
  str_fs <- list.files(here('_data/stressors_mol'),
                       pattern = smash2or(strs_unif),
                       full.names = TRUE)
  
  str_rast <- rast(str_fs) %>%
    setNames(basename(str_fs) %>% str_remove('_[0-9]{4}.+'))
  
  str_cell_df <- c(str_rast, cell_id_r) %>%
    as.data.frame(xy = FALSE) %>%
    filter(cell_id %in% cell_id_vec) %>%
    mutate(across(everything(), ~ifelse(is.na(.x), 0, .x))) %>%
    pivot_longer(cols = -cell_id, names_to = 'stressor', values_to = 'intensity')
  
  write_csv(str_cell_df, stressor_intensity_f)
}

str_cell_df <- read_csv(stressor_intensity_f)
unif_strs <- str_cell_df$stressor %>% unique()

chi_orig_spp_r <- rast(here('_output/cumulative_impact_maps/chi_species.tif')) %>%
  setNames('chi_orig_spp')
chi_orig_fe_r <- rast(here('_output/cumulative_impact_maps/chi_funct_entity.tif')) %>%
  setNames('chi_orig_fe')

chi_orig_df <- c(chi_orig_spp_r, chi_orig_fe_r, cell_id_r) %>%
  as.data.frame() %>%
  filter(cell_id %in% str_cell_df$cell_id)

```

```{r define function to calc impacts from vulns}
get_impacts <- function(f) {
  ### fs <- iter_df$f
  ### f <- fs[4]
  s <- str_remove_all(basename(f), '.+_shuffle_|_[0-9].+')
  if(!s %in% unif_strs) {
    df <- fread(f)
  } else {
    str_df <- str_cell_df %>%
      filter(stressor == s)
    if(nrow(str_df) < 100000) stop('missing rows for ', s)
    df <- fread(f) %>%
      oharac::dt_join(str_df, by = 'cell_id', type = 'inner') %>%
      mutate(imp_mean_spp = cell_vuln_spp * intensity,
             imp_mean_fe  = cell_vuln_fe  * intensity) %>%
      select(cell_id, imp_mean_spp, imp_mean_fe)
  }
  
  return(df)
}

calc_iter_chi <- function(i, tmp_df) {
  ### i <- 1
  message('Calculating mean chi for iteration ', i, '...')
  iter_df <- tmp_df %>%
    filter(iter == i)

  iter_cells_df <- parallel::mclapply(iter_df$f, FUN = get_impacts, mc.cores = 13) %>%
    setNames(iter_df$str) %>%
    rbindlist(idcol = 'stressor')
  
  iter_summary_df <- iter_cells_df %>%
    as.data.table() %>%
    .[ , .(chi_spp = sum(imp_mean_spp),
           chi_fe  = sum(imp_mean_fe)),
       by = cell_id] %>%
    mutate(i = i)
  
  return(iter_summary_df)
}
```

## Summarize impacts: CHI, non-CC, CC, and per stressor

```{r all impacts}

chi_out_f <- here_anx('spp_vs_fe_sensitivity_analysis/impact_summary/chi_shuffle_species.csv')

tmp_fs <- list.files(here_anx('spp_vs_fe_sensitivity_analysis/impact_summary/tmp'), full.names = TRUE)
tmp_df <- data.frame(f = tmp_fs) %>%
  mutate(iter  = str_extract(basename(f), '[0-9]+') %>% as.numeric(),
         str   = str_remove_all(basename(f), '.+_shuffle_|_[0-9].+'))

n_sims <- max(tmp_df$iter)

if(!file.exists(chi_out_f)) {
  ### unlink(chi_out_f)
  cell_chi_df <- lapply(1:n_sims, 
                        FUN = calc_iter_chi, 
                        tmp_df = tmp_df) %>%
    rbindlist() %>%
    oharac::dt_join(chi_orig_df, type = 'left', by = 'cell_id') %>%
    mutate(fe_spp_diff  = chi_fe - chi_spp,
           spp_spp_diff = chi_spp - chi_orig_spp,
           fe_fe_diff   = chi_fe - chi_orig_fe)
  
  cell_chi_summary <- cell_chi_df %>%
    as.data.table() %>%
    .[ , .(chi_mean_spp = mean(chi_spp),
           chi_sdev_spp = sd(chi_spp),
           chi_mean_fe = mean(chi_fe),
           chi_sdev_fe = sd(chi_fe),
           diff_95lo = quantile(fe_spp_diff, .025),
           diff_95hi = quantile(fe_spp_diff, .975),
           diff_95lo_spp = quantile(spp_spp_diff, .025),
           diff_95hi_spp = quantile(spp_spp_diff, .975),
           diff_95lo_fe  = quantile(fe_fe_diff, .025),
           diff_95hi_fe  = quantile(fe_fe_diff, .975),
           n_iters = length(chi_spp)),
       by = cell_id]
    
  write_csv(cell_chi_summary, chi_out_f)
}

```

```{r nonclimate impacts}

chi_noncc_out_f <- here_anx('spp_vs_fe_sensitivity_analysis/impact_summary/chi_noncc_shuffle_species.csv')

noncc_tmp_df <- tmp_df %>%
  filter(!str %in% c('sst_rise', 'ocean_acidification', 'sea_level_rise', 'sst_extremes', 'uv_radiation'))


noncc_orig_spp_r <- rast(here('_output/cumulative_impact_maps',
                              c('species_group_chi_fishing.tif', 
                                'species_group_chi_land_based.tif',
                                'species_group_chi_ocean_based.tif'))) %>%
  sum() %>%
  setNames('chi_orig_spp')
noncc_orig_fe_r <- rast(here('_output/cumulative_impact_maps',
                              c('funct_entity_group_chi_fishing.tif', 
                                'funct_entity_group_chi_land_based.tif',
                                'funct_entity_group_chi_ocean_based.tif'))) %>%
  sum() %>%
  setNames('chi_orig_fe')

noncc_orig_df <- c(noncc_orig_spp_r, noncc_orig_fe_r, cell_id_r) %>%
  as.data.frame() %>%
  filter(cell_id %in% str_cell_df$cell_id)

if(!file.exists(chi_noncc_out_f)) {
  ### unlink(chi_noncc_out_f)
  cell_chi_df <- lapply(1:n_sims, 
                        FUN = calc_iter_chi, 
                        tmp_df = noncc_tmp_df) %>%
    rbindlist() %>%
    oharac::dt_join(noncc_orig_df, type = 'left', by = 'cell_id') %>%
    mutate(fe_spp_diff  = chi_fe - chi_spp,
           spp_spp_diff = chi_spp - chi_orig_spp,
           fe_fe_diff   = chi_fe - chi_orig_fe)
  
  cell_chi_summary <- cell_chi_df %>%
    as.data.table() %>%
    .[ , .(chi_mean_spp = mean(chi_spp),
           chi_sdev_spp = sd(chi_spp),
           chi_mean_fe = mean(chi_fe),
           chi_sdev_fe = sd(chi_fe),
           diff_95lo = quantile(fe_spp_diff, .025),
           diff_95hi = quantile(fe_spp_diff, .975),
           diff_95lo_spp = quantile(spp_spp_diff, .025),
           diff_95hi_spp = quantile(spp_spp_diff, .975),
           diff_95lo_fe  = quantile(fe_fe_diff, .025),
           diff_95hi_fe  = quantile(fe_fe_diff, .975),
           n_iters = length(chi_spp)),
       by = cell_id]
    
  write_csv(cell_chi_summary, chi_noncc_out_f)
}

```

```{r climate impacts}

chi_cc_out_f <- here_anx('spp_vs_fe_sensitivity_analysis/impact_summary/chi_cc_shuffle_species.csv')

cc_tmp_df <- tmp_df %>%
  filter(str %in% c('sst_rise', 'ocean_acidification', 'sea_level_rise', 'sst_extremes', 'uv_radiation'))

cc_orig_spp_r <- rast(here('_output/cumulative_impact_maps',
                           'species_group_chi_climate.tif')) %>%
  sum() %>%
  setNames('chi_orig_spp')
cc_orig_fe_r <- rast(here('_output/cumulative_impact_maps',
                          'funct_entity_group_chi_climate.tif')) %>%
  sum() %>%
  setNames('chi_orig_fe')

noncc_orig_df <- c(noncc_orig_spp_r, noncc_orig_fe_r, cell_id_r) %>%
  as.data.frame() %>%
  filter(cell_id %in% str_cell_df$cell_id)


if(!file.exists(chi_cc_out_f)) {
  ### unlink(chi_cc_out_f)

  cell_chi_df <- lapply(1:n_sims, 
                        FUN = calc_iter_chi, 
                        tmp_df = cc_tmp_df) %>%
    rbindlist() %>%
    oharac::dt_join(chi_orig_df, type = 'left', by = 'cell_id') %>%
    mutate(fe_spp_diff  = chi_fe - chi_spp,
           spp_spp_diff = chi_spp - chi_orig_spp,
           fe_fe_diff   = chi_fe - chi_orig_fe)
  
  cell_chi_summary <- cell_chi_df %>%
    as.data.table() %>%
    .[ , .(chi_mean_spp = mean(chi_spp),
           chi_sdev_spp = sd(chi_spp),
           chi_mean_fe = mean(chi_fe),
           chi_sdev_fe = sd(chi_fe),
           diff_95lo = quantile(fe_spp_diff, .025),
           diff_95hi = quantile(fe_spp_diff, .975),
           diff_95lo_spp = quantile(spp_spp_diff, .025),
           diff_95hi_spp = quantile(spp_spp_diff, .975),
           diff_95lo_fe  = quantile(fe_fe_diff, .025),
           diff_95hi_fe  = quantile(fe_fe_diff, .975),
           n_iters = length(chi_spp)),
       by = cell_id]
    
  write_csv(cell_chi_summary, chi_cc_out_f)
}

```

```{r per-stressor impacts}

impact_out_fstem <- here_anx('spp_vs_fe_sensitivity_analysis/impact_summary/impact_%s_shuffle_species.csv')
### unlink_fs <- list.files(dirname(impact_out_fstem), pattern = 'species.csv', full.names = TRUE)
### unlink(unlink_fs)

strs <- tmp_df$str %>% unique()

for(s in strs) {
  # s <- strs[1]
  str_tmp_df <- tmp_df %>%
    filter(str == s)

  impact_out_f <- sprintf(impact_out_fstem, s)
  if(!file.exists(impact_out_f)) {
    cell_imp_df <- lapply(1:n_sims, 
                          FUN = calc_iter_chi, 
                          tmp_df = str_tmp_df) %>%
      rbindlist() %>%
      setNames(str_replace(names(.), 'chi', 'impact')) %>%
      mutate(fe_spp_diff = impact_fe - impact_spp)
    
    cell_impact_summary <- cell_imp_df %>%
      as.data.table() %>%
      .[ , .(impact_mean_spp = mean(impact_spp),
             impact_sdev_spp = sd(impact_spp),
             impact_mean_fe = mean(impact_fe),
             impact_sdev_fe = sd(impact_fe),
             diff_95lo = quantile(fe_spp_diff, .025),
             diff_95hi = quantile(fe_spp_diff, .975),
             n_iters = length(impact_spp)),
         by = cell_id]
      
    write_csv(cell_impact_summary, impact_out_f)
  }
}




```
